% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtsSphere-package.R
\docType{package}
\name{mrtsSphere-package}
\alias{mrtsSphere}
\alias{mrtsSphere-package}
\title{mrtsSphere: Multi-Resolution Thin-Plate Splines on the Sphere}
\description{
Constructs multi-resolution thin-plate spline basis functions on the
sphere for use in spatial regression and large-scale spatial prediction
problems.
}
\details{
The main user-facing function is \code{\link[=mrts_sphere]{mrts_sphere()}}.
}
\references{
Multi-resolution approximations of Gaussian processes for large
spatial datasets on the sphere. \emph{Environmetrics}, 2025.
\doi{10.1002/env.70092}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/STLABTW/multi-resolution-sphere}
  \item Report bugs at \url{https://github.com/STLABTW/multi-resolution-sphere/issues}
}

}
\author{
\strong{Maintainer}: STLAB Authors \email{hhuscout@gmail.com}

}
\keyword{internal}
